requirements.txt
streamlit>=1.28.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.2
langchain-chroma>=0.0.1
langchain-groq>=0.1.1
langchain-classic>=0.0.11
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pypdf>=3.17.0
chromadb>=0.4.15
unstructured>=0.10.30
pdf2image>=1.16.3
pytesseract>=0.3.10